requirements.txt
python-telegram-bot==20.8
apscheduler==3.10.4
googletrans==4.0.0-rc1
